modualizied config files and scripts for interacting with smart contracts.
