+++ rightmesh/config/general-settings.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // geth instance that the scripts should interact with
+  rpcHost: 'http://localhost:8545',
+}
