+++ rightmesh/config/token-config.js
@@ -1,0 +1,8 @@
+const utils = require('../utils/utils');
+
+module.exports = {
+  // account from geth that is being used for deployment
+  // this account will also be the owner of crowdsale and eventually token contract
+  ownerAccount: '0xa30109d16db929bb45e08db7127d92a5f5114472',
+  tokenAddress: '0x0E32C1A5C1F8873f67eAE883fF8D228A713D5C72',
+}
